Application works with multi client connections, can be done across different devices, different internet connections, all connect to server and exchange messaged.
Also added new chat button implementation, as well as my ip display and copy of ip on click. The ip is supposed to function as a code that you would give your contacts, so that they can create a chat session with you.

Issues:
- Server program has some kind of threading issue overtime (seems to be client connections mostly, builds resource usage overtime but fairly quickly if client stays connected).

- Disconnections from one instance of PRISM on when 2 are running on same computer and reconnects seems to cause problems. Not a major issue if it's only caused by multiple instances on the same machine

Next to do:
Keep implementing the chat with desired contacts feature, then after that if possible do:
- Group Chats
- Account logs